# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Cart API",
        "title": "Cart",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/cart": {
            "post": {
                "description": "Write order to redis",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Write order",
                "parameters": [
                    {
                        "description": "body",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cart.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "503": {
                        "description": "Service Unavailable"
                    }
                }
            }
        },
        "/cart/{key}": {
            "get": {
                "description": "Read order from redis",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Read order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cart.Order"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "503": {
                        "description": "Service Unavailable"
                    }
                }
            }
        }
    },
    "definitions": {
        "cart.Order": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}